Correccion de componente ApartadoClient

+++ src/pages/ApartadosClient/ApartadosClient.jsx
@@ -6,6 +6,7 @@ import Footer from '../../components/Footer/Footer';
 import Loader from '../../components/Loader/Loader';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
+import Apartados from '../Apartados/Apartados';
 
 export default function ApartadosClient() {
   const [apartados, setApartados] = useState([]);
@@ -16,13 +17,14 @@ export default function ApartadosClient() {
     const fetchApartados = async () => {
       try {
         const token = localStorage.getItem('token');
-        const response = await fetch('http://localhost:8080/v1/Apartados?page=1&size=10', {
+        const response = await fetch('http://localhost:8080/v1/Apartados?page=1&size=100', {
           headers: {
             Authorization: `Bearer ${token}`,
           },
         });
         const data = await response.json();
         setApartados(data.data);
+        console.log(data.data)
       } catch (error) {
         console.error('Error fetching apartados:', error);
       } finally {
@@ -49,9 +51,15 @@ export default function ApartadosClient() {
         },
       });
       const productDetails = await response.json();
-      console.log(productDetails.data[0]);
-      return productDetails.data[0];
-     
+
+      // Verificar si hay datos y si hay un objeto en la posición 0 antes de acceder a las propiedades
+      if (productDetails.data && productDetails.data.length > 0) {
+        console.log('productdetails', productDetails.data[0]);
+        return productDetails.data[0];
+      } else {
+        console.error('Error fetching product details: No data found');
+        return null;
+      }
     } catch (error) {
       console.error('Error fetching product details:', error);
       return null;
@@ -75,7 +83,6 @@ export default function ApartadosClient() {
       setTimeout(() => {
         window.location.reload();
       }, 2000);
-     
     } catch (error) {
       console.error('Error deleting apartado:', error);
     }
@@ -83,70 +90,75 @@ export default function ApartadosClient() {
 
   const renderCards = async () => {
     const clienteApartados = apartados.filter((apartado) => apartado.clienteId === parseInt(clienteId));
-  
+
     const cards = await Promise.all(
       clienteApartados.map(async (apartado) => {
         const productDetails = await renderCardDetails(apartado.productoId);
-  
-        // Calcular la fecha de vigencia y la fecha de notificación
-        const fechaVigencia = new Date(apartado.createdAt);
-        fechaVigencia.setDate(fechaVigencia.getDate() + 3); // Sumar 3 días
-  
-        const fechaNotificacion = new Date(fechaVigencia);
-        fechaNotificacion.setDate(fechaNotificacion.getDate() - 1); // Restar 1 día para notificar un día antes
-  
-        // Calcular el tiempo restante hasta la notificación en días
-        const tiempoRestante = Math.ceil((fechaNotificacion - Date.now()) / (1000 * 60 * 60 * 24)); // Convertir de milisegundos a días
-  
-        // Texto y estilo para mostrar en la interfaz
-        let vigenciaText = `Vigencia: ${tiempoRestante} días`;
-        let vigenciaStyle = { color: 'black', display:'flex', alignItems:'Center', justifyContent:'center', width:'100%', margin:'auto' }; // Color por defecto
-  
-        // Si la vigencia es 0 días, cambiar el texto a rojo y mostrar "Vence hoy"
-        if (tiempoRestante === 0) {
-          vigenciaText = 'Vence hoy';
-          vigenciaStyle = { color: 'red' , display:'flex', alignItems:'Center', justifyContent:'center', width:'100%', margin:'auto'};
-        }
-        if (tiempoRestante < 0) {
-            vigenciaText = 'Apartado Vencido';
-            vigenciaStyle = { color: 'red' , display:'flex', alignItems:'Center', justifyContent:'center', width:'100%', margin:'auto'};
+
+        // Verificar si productDetails es null antes de intentar acceder a las propiedades
+        if (productDetails) {
+          // Calcular la fecha de vigencia y la fecha de notificación
+          const fechaVigencia = new Date(apartado.createdAt);
+          fechaVigencia.setDate(fechaVigencia.getDate() + 3); // Sumar 3 días
+
+          const fechaNotificacion = new Date(fechaVigencia);
+          fechaNotificacion.setDate(fechaNotificacion.getDate() - 1); // Restar 1 día para notificar un día antes
+
+          // Calcular el tiempo restante hasta la notificación en días
+          const tiempoRestante = Math.ceil((fechaNotificacion - Date.now()) / (1000 * 60 * 60 * 24)); // Convertir de milisegundos a días
+
+          // Texto y estilo para mostrar en la interfaz
+          let vigenciaText = `Vigencia: ${tiempoRestante} días`;
+          let vigenciaStyle = { color: 'black', display: 'flex', alignItems: 'Center', justifyContent: 'center', width: '100%', margin: 'auto' }; // Color por defecto
+
+          // Si la vigencia es 0 días, cambiar el texto a rojo y mostrar "Vence hoy"
+          if (tiempoRestante === 0) {
+            vigenciaText = 'Vence hoy';
+            vigenciaStyle = { color: 'red', display: 'flex', alignItems: 'Center', justifyContent: 'center', width: '100%', margin: 'auto' };
           }
-        // Programar la notificación
-        setTimeout(() => {
-          // Notificar al cliente
-          console.log(`Notificar al cliente: El apartado ${apartado.apartadoId} está a punto de vencer.`);
-        }, tiempoRestante * 24 * 60 * 60 * 1000); // Convertir días a milisegundos
-  
-        // Programar la eliminación del apartado
-        setTimeout(() => {
-          deleteApartado(apartado.apartadoId);
-        }, fechaVigencia - Date.now());
-  
-        return (
-          <>
-            <div style={{ pointerEvents: 'none', display: 'flex', flexDirection: 'column', margin: 'auto' }}>
+          if (tiempoRestante < 0) {
+            vigenciaText = 'Apartado Vencido';
+            vigenciaStyle = { color: 'red', display: 'flex', alignItems: 'Center', justifyContent: 'center', width: '100%', margin: 'auto' };
+          }
+          // Programar la notificación
+          setTimeout(() => {
+            // Notificar al cliente
+            console.log(`Notificar al cliente: El apartado ${apartado.apartadoId} está a punto de vencer.`);
+          }, tiempoRestante * 24 * 60 * 60 * 1000); // Convertir días a milisegundos
+
+          // Programar la eliminación del apartado
+          setTimeout(() => {
+            deleteApartado(apartado.apartadoId);
+          }, fechaVigencia - Date.now());
+
+          return (
+            <div key={apartado.apartadoId} style={{ pointerEvents: 'none', display: 'flex', flexDirection: 'column', margin: 'auto' }}>
               <p style={{ fontSize: '0.9rem', ...vigenciaStyle }}>{vigenciaText}</p>
+              <p style={{ fontSize: '0.9rem', ...vigenciaStyle }}>Numero de apartado: <p style={{backgroundColor:'skyblue', borderRadius:'10px'}}>{apartado.apartadoId}</p></p>
+
               <Card
                 key={apartado.apartadoId}
-                id={productDetails.productoId}
+                id={apartado.apartadoId}
                 modelo={productDetails.modelo}
                 cantidad={apartado.cantidad}
                 precio={productDetails.precio}
                 subTotal={apartado.subTotal}
-              
                 descuento={apartado.descuento}
                 total={apartado.total}
                 imagen={productDetails ? productDetails.url_calzado : ''}
               />
             </div>
-          </>
-        );
+          );
+        } else {
+          // Manejar el caso donde no se pueden obtener los detalles del producto
+          console.error('Error rendering card: Product details not available');
+          return null;
+        }
       })
     );
-  
-    setRenderedCards(cards);
+
+    setRenderedCards(cards.filter(card => card !== null));
   };
-  
 
   useEffect(() => {
     if (!loading) {
@@ -162,7 +174,7 @@ export default function ApartadosClient() {
       <div className={style.contenido}>
         {loading ? (
           <>
-            <div style={{ display: 'flex', flexDirection: 'column', width:'90%', height: '48vh', justifyContent: 'center', alignItems: 'center' }}>
+            <div style={{ display: 'flex', flexDirection: 'column', width: '90%', height: '48vh', justifyContent: 'center', alignItems: 'center' }}>
               <Loader />
               <p>Loading...</p>
             </div>
